requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
websockets
python-dotenv